DABOM-292 chore: recap 엔티티 디렉토리 이동
